3.1.1 - Skopiowanie Implementacji do IntArrayStack
